Main page changes
