+++ flag.txt
@@ -1,3 +1,3 @@
 True
-True
-True
+False
+False
